pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bods-brightquery"
version = "0.1.0"
description = "Transform BrightQuery/OpenData.org data into Beneficial Ownership Data Standard (BODS) v0.4 format"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Stephen Abbott Pugh"},
]
dependencies = [
    "click>=8.1",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
bods-bq = "bods_brightquery.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
